refactor project visibility into it's own component, and add tests
